chore(crypto): move test scripts to own files
